requirements.txt
Flask
psycopg2-binary
boto3
bcrypt
